+++ mocket/__init__.py
@@ -1,7 +1,8 @@
 from mocket.async_mocket import async_mocketize
 from mocket.entry import MocketEntry
-from mocket.mocket import FakeSSLContext, Mocket
+from mocket.mocket import Mocket
 from mocket.mocketizer import Mocketizer, mocketize
+from mocket.ssl import FakeSSLContext
 
 __all__ = (
     "async_mocketize",
